+++ metadata.rb
@@ -2,6 +2,6 @@ name             'siebel'
 maintainer       'YOUR_COMPANY_NAME'
 maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'
-description      'Installs/Configures siebelcrm'
+description      'Installs/Configures siebel'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'