+++ recipes/install_swes.rb
@@ -1,12 +1,12 @@
-swes_install_folder = "#{node[:siebelcrm][:distr][:folder]}/swes/Disk1"
+swes_install_folder = "#{node[:siebel][:distr][:folder]}/swes/Disk1"
 
-unless File.exists?(node[:siebelcrm][:ses_home][:location])
+unless File.exists?(node[:siebel][:ses_home][:location])
   unless File.exists?(swes_install_folder) 
-    include_recipe 'siebelcrm::prepare_distrib_for_swes' 
+    include_recipe 'siebel::prepare_distrib_for_swes' 
   end
 
   # Creating SIEBEL_HOME 
-  directory "#{node[:siebelcrm][:swes_home][:location]}" do
+  directory "#{node[:siebel][:swes_home][:location]}" do
     recursive true
     owner "oracle"
     group "oinstall"