+++ recipes/prepare_folders.rb
@@ -1,5 +1,5 @@
 # Creating SIEBEL_HOME 
-directory "#{node[:siebelcrm][:oracle_home][:location]}" do
+directory "#{node[:siebel][:oracle_home][:location]}" do
   recursive true
   owner "oracle"
   group "oinstall"