+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-siebelcrm CHANGELOG
+siebel CHANGELOG
 ===================
 
-This file is used to list changes made in each version of the siebelcrm cookbook.
+This file is used to list changes made in each version of the siebel cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of siebelcrm
+- [your_name] - Initial release of siebel
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.